Refactor code to separate resource objects from resource hierarchy objects
